requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.5
djangorestframework==3.14.0
flake8==6.1.0
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
